Reorganize test files to follow new structure
