+++ tests/unit/objects/test_delete.py
