+++ tests/unit/objects/test_write.py
